+++ src/lyrics_fetcher.h
@@ -118,7 +118,7 @@ struct GeniusFetcher : public GoogleLyricsFetcher
 	virtual const char *name() const override { return "genius.com"; }
 
 protected:
-	virtual const char *regex() const override { return "<div class=\"Lyrics__Container.*?\">(.*?)</div>"; }
+	virtual const char *regex() const override { return "<div class=\"[Ll]yrics.*?>(.*?)</div>"; }
 };
 
 struct JahLyricsFetcher : public GoogleLyricsFetcher